added telemetry tracking for quickfixes on hover, fixed import cycle error, and cleaned up code
