--- src/vs/editor/contrib/codeAction/browser/codeMenuOpenedFrom.ts
@@ -1,15 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-export enum CodeMenuOpenedFrom {
-	Refactor = 'from: refactor',
-	RefactorPreview = 'from: refactor preview',
-	Lightbulb = 'from: lightbulb',
-	Default = 'from: other (default)',
-	SourceAction = 'from: source action',
-	QuickFix = 'from: quick fix',
-	FixAll = 'from: fix all',
-	OrganizeImports = 'from: organize imports',
-	AutoFix = 'from: auto fix'
-}
